feat(data): update JDBC DAO implementations with DatabaseConfig

+++ auction-server/src/main/java/com/auction/server/dao/impl/JdbcAuctionDao.java
@@ -13,7 +13,7 @@ import java.util.Optional;
 /**
  * Triển khai AuctionDao sử dụng JDBC + MySQL.
  *
- * <p>Pessimistic Locking: Phương thức findById dùng SELECT ... FOR UPDATE để khóa dòng dữ liệu khi
+ * Pessimistic Locking: Phương thức findById dùng SELECT ... FOR UPDATE để khóa dòng dữ liệu khi
  * AuctionService đang xử lý lệnh đặt giá (bid). Nếu 2 luồng cùng cố gắng đặt giá trên cùng 1
  * Auction, luồng thứ 2 bị Database ĐẬP CỬA giữ lại cho đến khi luồng thứ 1 COMMIT xong. Đây là cơ
  * chế chống Lost Update ở tầng Database, đảm bảo tính Concurrent Bidding của hệ thống.
@@ -36,16 +36,14 @@ public class JdbcAuctionDao implements AuctionDao {
 
   @Override
   public Auction save(Auction auction) {
-    String sql =
-        "INSERT INTO auctions (created_at, item_id, seller_id, current_price,"
-            + " current_winner_id, status, start_time, end_time)"
-            + " VALUES (?, ?, ?, ?, ?, ?, ?, ?)";
+    String sql = "INSERT INTO auctions (created_at, item_id, seller_id, current_price,"
+        + " current_winner_id, status, start_time, end_time)" + " VALUES (?, ?, ?, ?, ?, ?, ?, ?)";
 
     try (PreparedStatement ps = conn.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
       ps.setTimestamp(1, Timestamp.valueOf(auction.getCreatedAt()));
       ps.setLong(2, auction.getItemId());
       ps.setLong(3, auction.getSellerId());
-      ps.setDouble(4, auction.getCurrentPrice());
+      ps.setLong(4, auction.getCurrentPrice());
       setNullableLong(ps, 5, auction.getCurrentWinnerId()); // null nếu chưa có ai đặt giá
       ps.setString(6, auction.getStatus().name());
       ps.setTimestamp(7, Timestamp.valueOf(auction.getStartTime()));
@@ -53,7 +51,8 @@ public class JdbcAuctionDao implements AuctionDao {
       ps.executeUpdate();
 
       try (ResultSet keys = ps.getGeneratedKeys()) {
-        if (keys.next()) auction.setId(keys.getLong(1));
+        if (keys.next())
+          auction.setId(keys.getLong(1));
       }
       return auction;
     } catch (SQLException e) {
@@ -64,7 +63,8 @@ public class JdbcAuctionDao implements AuctionDao {
   /**
    * Tìm Auction theo ID với khóa Pessimistic (FOR UPDATE).
    *
-   * <p>Khi AuctionService cần đặt giá, nó gọi hàm này trong 1 transaction. H2/MySQL sẽ khóa dòng
+   * <p>
+   * Khi AuctionService cần đặt giá, nó gọi hàm này trong 1 transaction. H2/MySQL sẽ khóa dòng
    * auction đó lại. Mọi luồng khác gọi findById cùng ID sẽ bị block cho đến khi transaction kết
    * thúc (COMMIT/ROLLBACK). Điều này tuyệt đối đảm bảo chỉ 1 luồng cập nhật giá tại 1 thời điểm.
    */
@@ -75,7 +75,8 @@ public class JdbcAuctionDao implements AuctionDao {
     try (PreparedStatement ps = conn.prepareStatement(sql)) {
       ps.setLong(1, id);
       try (ResultSet rs = ps.executeQuery()) {
-        if (rs.next()) return Optional.of(mapRow(rs));
+        if (rs.next())
+          return Optional.of(mapRow(rs));
       }
     } catch (SQLException e) {
       throw new AuctionException("Database error finding Auction by id", e);
@@ -87,9 +88,9 @@ public class JdbcAuctionDao implements AuctionDao {
   public List<Auction> findAll() {
     List<Auction> list = new ArrayList<>();
     String sql = "SELECT * FROM auctions";
-    try (Statement st = conn.createStatement();
-        ResultSet rs = st.executeQuery(sql)) {
-      while (rs.next()) list.add(mapRow(rs));
+    try (Statement st = conn.createStatement(); ResultSet rs = st.executeQuery(sql)) {
+      while (rs.next())
+        list.add(mapRow(rs));
     } catch (SQLException e) {
       throw new AuctionException("Database error fetching all Auctions", e);
     }
@@ -98,11 +99,10 @@ public class JdbcAuctionDao implements AuctionDao {
 
   @Override
   public Auction update(Auction auction) {
-    String sql =
-        "UPDATE auctions SET current_price=?, current_winner_id=?, status=?, end_time=?"
-            + " WHERE id=?";
+    String sql = "UPDATE auctions SET current_price=?, current_winner_id=?, status=?, end_time=?"
+        + " WHERE id=?";
     try (PreparedStatement ps = conn.prepareStatement(sql)) {
-      ps.setDouble(1, auction.getCurrentPrice());
+      ps.setLong(1, auction.getCurrentPrice());
       setNullableLong(ps, 2, auction.getCurrentWinnerId());
       ps.setString(3, auction.getStatus().name());
       ps.setTimestamp(4, Timestamp.valueOf(auction.getEndTime()));
@@ -141,8 +141,7 @@ public class JdbcAuctionDao implements AuctionDao {
   @Override
   public long count() {
     String sql = "SELECT COUNT(1) FROM auctions";
-    try (Statement st = conn.createStatement();
-        ResultSet rs = st.executeQuery(sql)) {
+    try (Statement st = conn.createStatement(); ResultSet rs = st.executeQuery(sql)) {
       return rs.next() ? rs.getLong(1) : 0L;
     } catch (SQLException e) {
       throw new AuctionException("Database error counting Auctions", e);
@@ -160,7 +159,8 @@ public class JdbcAuctionDao implements AuctionDao {
     try (PreparedStatement ps = conn.prepareStatement(sql)) {
       ps.setString(1, status.name());
       try (ResultSet rs = ps.executeQuery()) {
-        while (rs.next()) list.add(mapRow(rs));
+        while (rs.next())
+          list.add(mapRow(rs));
       }
     } catch (SQLException e) {
       throw new AuctionException("Database error finding Auctions by status", e);
@@ -174,7 +174,8 @@ public class JdbcAuctionDao implements AuctionDao {
     try (PreparedStatement ps = conn.prepareStatement(sql)) {
       ps.setLong(1, itemId);
       try (ResultSet rs = ps.executeQuery()) {
-        if (rs.next()) return Optional.of(mapRow(rs));
+        if (rs.next())
+          return Optional.of(mapRow(rs));
       }
     } catch (SQLException e) {
       throw new AuctionException("Database error finding Auction by item ID", e);
@@ -189,7 +190,8 @@ public class JdbcAuctionDao implements AuctionDao {
     try (PreparedStatement ps = conn.prepareStatement(sql)) {
       ps.setLong(1, sellerId);
       try (ResultSet rs = ps.executeQuery()) {
-        while (rs.next()) list.add(mapRow(rs));
+        while (rs.next())
+          list.add(mapRow(rs));
       }
     } catch (SQLException e) {
       throw new AuctionException("Database error finding Auctions by seller ID", e);
@@ -206,9 +208,9 @@ public class JdbcAuctionDao implements AuctionDao {
   public List<Auction> findActiveAuctions() {
     List<Auction> list = new ArrayList<>();
     String sql = "SELECT * FROM auctions WHERE status IN ('OPEN', 'RUNNING')";
-    try (Statement st = conn.createStatement();
-        ResultSet rs = st.executeQuery(sql)) {
-      while (rs.next()) list.add(mapRow(rs));
+    try (Statement st = conn.createStatement(); ResultSet rs = st.executeQuery(sql)) {
+      while (rs.next())
+        list.add(mapRow(rs));
     } catch (SQLException e) {
       throw new AuctionException("Database error finding active Auctions", e);
     }
@@ -225,14 +227,9 @@ public class JdbcAuctionDao implements AuctionDao {
       currentWinnerId = null;
     }
 
-    return new Auction(
-        rs.getLong("id"),
-        rs.getTimestamp("created_at").toLocalDateTime(),
-        rs.getLong("item_id"),
-        rs.getLong("seller_id"),
-        rs.getDouble("current_price"),
-        currentWinnerId,
-        AuctionStatus.valueOf(rs.getString("status")),
+    return new Auction(rs.getLong("id"), rs.getTimestamp("created_at").toLocalDateTime(),
+        rs.getLong("item_id"), rs.getLong("seller_id"), rs.getLong("current_price"),
+        currentWinnerId, AuctionStatus.valueOf(rs.getString("status")),
         rs.getTimestamp("start_time").toLocalDateTime(),
         rs.getTimestamp("end_time").toLocalDateTime());
   }
